UIMA-4577 First delivery, DB project.
git-svn-id: https://svn.apache.org/repos/asf/uima/sandbox/uima-ducc/trunk@1703203 13f79535-47bb-0310-9956-ffa450edef68

+++ uima-ducc-common/src/main/java/org/apache/uima/ducc/common/persistence/services/StateServicesDirectory.java
@@ -23,17 +23,17 @@ import java.util.TreeMap;
 
 public class StateServicesDirectory {
 
-	private TreeMap<Integer,StateServicesSet> map = new TreeMap<Integer,StateServicesSet>();
+	private TreeMap<Long,StateServicesSet> map = new TreeMap<Long,StateServicesSet>();
 	
-	public void put(Integer key, StateServicesSet value) {
+	public void put(Long key, StateServicesSet value) {
 		map.put(key, value);
 	}
 	
-	public StateServicesSet get(Integer key) {
+	public StateServicesSet get(Long key) {
 		return map.get(key);
 	}
 	
-	public NavigableSet<Integer> getDescendingKeySet() {
+	public NavigableSet<Long> getDescendingKeySet() {
 		return map.descendingKeySet();
 	}
 }
